Extract MessagePadding into a separate file
